+++ conf/evolutions/default/1.sql
@@ -40,7 +40,7 @@ create table game (
   user_team_id              bigint,
   is_team_init              boolean,
   is_roster_init            boolean,
-  money                     integer,
+  money                     bigint,
   date                      bigint,
   constraint pk_game primary key (id))
 ;
